+++ src/contents/ui/Week.qml
@@ -17,7 +17,7 @@ ColumnLayout {
 
     // ------------------------------------------------------------------------------------------------------------------------
 
-	// we always count from 0 being sunday
+	// we always count from 0 being sunday unless user set different day to be first day of the week
 	property int weekdayOffset: 0
 	property int firstDayOfWeek: 0
 
@@ -25,23 +25,39 @@ ColumnLayout {
 
 	// https://api.kde.org/frameworks/plasma-framework/html/classPlasma_1_1QuickTheme.html
 	// or PlasmaCore.Theme.* (starting frameworks 5.73)
-	property string pastDayFg: ccEnabled ? plasmoid.configuration.customColorsPastDayFg : theme.disabledTextColor
-	property string pastDayBg: ccEnabled ? plasmoid.configuration.customColorsPastDayBg : theme.backgroundColor
-	property bool pastDayBold: ccEnabled ? plasmoid.configuration.customColorsPastDayBold: false
-	property bool pastDayItalic: ccEnabled ? plasmoid.configuration.customColorsPastDayItalic: false
+	readonly property string pastDayFg: ccEnabled ? plasmoid.configuration.customColorsPastDayFg : theme.disabledTextColor
+	readonly property string pastDayBg: ccEnabled ? plasmoid.configuration.customColorsPastDayBg : theme.backgroundColor
+	readonly property bool pastDayBold: ccEnabled ? plasmoid.configuration.customColorsPastDayBold : false
+	readonly property bool pastDayItalic: ccEnabled ? plasmoid.configuration.customColorsPastDayItalic : false
 
-	property string todayFg: ccEnabled ? plasmoid.configuration.customColorsTodayFg : "#FFffffff"
-	property string todayBg: ccEnabled ? plasmoid.configuration.customColorsTodayBg : "#FFff006e"
-	property bool todayBold: ccEnabled ? plasmoid.configuration.customColorsTodayBold: true
-	property bool todayItalic: ccEnabled ? plasmoid.configuration.customColorsTodayItalic: false
+	readonly property string todayFg: ccEnabled ? plasmoid.configuration.customColorsTodayFg : "#FFffffff"
+	readonly property string todayBg: ccEnabled ? plasmoid.configuration.customColorsTodayBg : "#FFff006e"
+	readonly property bool todayBold: ccEnabled ? plasmoid.configuration.customColorsTodayBold : true
+	readonly property bool todayItalic: ccEnabled ? plasmoid.configuration.customColorsTodayItalic : false
 
-	property string futureDayFg: ccEnabled ? plasmoid.configuration.customColorsFutureDayFg : theme.textColor
-	property string futureDayBg: ccEnabled ? plasmoid.configuration.customColorsFutureDayBg : theme.backgroundColor
-	property bool futureDayBold: ccEnabled ? plasmoid.configuration.customColorsFutureDayBold: false
-	property bool futureDayItalic: ccEnabled ? plasmoid.configuration.customColorsFutureDayItalic: false
+	readonly property string futureDayFg: ccEnabled ? plasmoid.configuration.customColorsFutureDayFg : theme.textColor
+	readonly property string futureDayBg: ccEnabled ? plasmoid.configuration.customColorsFutureDayBg : theme.backgroundColor
+	readonly property bool futureDayBold: ccEnabled ? plasmoid.configuration.customColorsFutureDayBold : false
+	readonly property bool futureDayItalic: ccEnabled ? plasmoid.configuration.customColorsFutureDayItalic : false
+
+	readonly property bool ccSaturdayEnabled: plasmoid.configuration.customColorsFutureSaturdayEnabled
+	readonly property string futureSaturdayFg: ccSaturdayEnabled ? plasmoid.configuration.customColorsFutureSaturdayFg : futureDayFg
+	readonly property string futureSaturdayBg: ccSaturdayEnabled ? plasmoid.configuration.customColorsFutureSaturdayBg : futureDayBg
+	readonly property bool futureSaturdayBold: ccSaturdayEnabled ? plasmoid.configuration.customColorsFutureSaturdayBold : futureDayBold
+	readonly property bool futureSaturdayItalic: ccSaturdayEnabled ? plasmoid.configuration.customColorsFutureSaturdayItalic : futureDayItalic
+
+	readonly property bool ccSundayEnabled: plasmoid.configuration.customColorsFutureSundayEnabled
+	readonly property string futureSundayFg: ccSundayEnabled ? plasmoid.configuration.customColorsFutureSundayFg : futureDayFg
+	readonly property string futureSundayBg: ccSundayEnabled ? plasmoid.configuration.customColorsFutureSundayBg : futureDayBg
+	readonly property bool futureSundayBold: ccSundayEnabled ? plasmoid.configuration.customColorsFutureSundayBold : futureDayBold
+	readonly property bool futureSundayItalic: ccSundayEnabled ? plasmoid.configuration.customColorsFutureSundayItalic : futureDayItalic
 
 	// we start from Sunday here
 	property var labels: ['', '', '', '', '', '', '']
+	property var attrsBg: ['#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff']
+	property var attrsFg: ['#00000000', '#00000000', '#00000000', '#00000000', '#00000000', '#00000000', '#00000000']
+	property var attrsBold: [false, false, false, false, false, false, false]
+	property var attrsItalic: [false, false, false, false, false, false, false]
 
     // ------------------------------------------------------------------------------------------------------------------------
 
@@ -69,10 +85,6 @@ ColumnLayout {
 				firstDayOfWeek = locale.firstDayOfWeek
 			}
 
-			var now = new Date()
-			weekdayOffset = now.getDay()-firstDayOfWeek
-			if (weekdayOffset < 0) weekdayOffset += 7
-
 			// it's known that Jan 5, 2020 is Sunday
 			// so we shift the index offset (based on locale) so the first
 			// cell of 'labels' is always first day of the week.
@@ -80,14 +92,82 @@ ColumnLayout {
 
 			for(var i=0; i<7; i++) {
 				// it's known that Jan 5, 2020 is Sunday
-				var day = new Date(2020, 0, 5+firstDayOfWeek+i)
+				var day = new Date(2020, 0, 5 + (firstDayOfWeek + i))
 				tmp[i] = day.toLocaleDateString(locale, 'ddd').substr(0, 1).toUpperCase()
 			}
 			labels = tmp
+
+			// fg colors
+			var fgTmp = ['#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff', '#FFffffff']
+			var bgTmp = ['#00000000', '#00000000', '#00000000', '#00000000', '#00000000', '#00000000', '#00000000']
+			var boldTmp = [false, false, false, false, false, false, false]
+			var italicTmp = [false, false, false, false, false, false, false]
+
+			var now = new Date()
+			weekdayOffset = now.getDay()-firstDayOfWeek
+			if (weekdayOffset < 0) weekdayOffset += 7
+
+			// Where's Today's cell in our week grid?
+			var todayOffset = now.getDay() - firstDayOfWeek
+			if (todayOffset < 0) todayOffset += 7
+	
+			for(var i=0; i<7; i++) {
+				var weekday = (i+firstDayOfWeek) % 7
+				fgTmp[i] = getFg(i, weekday, todayOffset)
+				bgTmp[i] = getBg(i, weekday, todayOffset)
+				boldTmp[i] = getBold(i, weekday, todayOffset)
+				italicTmp[i] = getItalic(i, weekday, todayOffset)
+			}
+
+			attrsFg = fgTmp
+			attrsBg = bgTmp
+			attrsBold = boldTmp
+			attrsItalic = italicTmp
 		}
 	}
 
     // ------------------------------------------------------------------------------------------------------------------------
+
+	function getFg(index, weekday, todayOffset) {
+		if (index === todayOffset) return todayFg
+		if (index < todayOffset) return pastDayFg
+		var result = futureDayFg
+		switch(weekday) {
+			case 0: result = futureSundayFg; break
+			case 6: result = futureSaturdayFg; break
+		}
+		return result
+	}
+	function getBg(index, weekday, todayOffset) {
+		if (index === todayOffset) return todayBg
+		if (index < todayOffset) return pastDayBg
+		var result = futureDayBg
+		switch(weekday) {
+			case 0: result = futureSundayBg; break
+			case 6: result = futureSaturdayBg; break
+		}
+		return result
+	}
+	function getBold(index, weekday, todayOffset) {
+		if (index === todayOffset) return todayBold
+		if (index < todayOffset) return pastDayBold
+		var result = futureDayBold
+		switch(weekday) {
+			case 0: result = futureSundayBold; break
+			case 6: result = futureSaturdayBold; break
+		}
+		return result
+	}
+	function getItalic(index, weekday, todayOffset) {
+		if (index === todayOffset) return todayItalic
+		if (index < todayOffset) return pastDayItalic
+		var result = futureDayItalic
+		switch(weekday) {
+			case 0: result = futureSundayItalic; break
+			case 6: result = futureSaturdayItalic; break
+		}
+		return result
+	}
 
 	GridLayout {
 		id: weekGrid
@@ -104,22 +184,10 @@ ColumnLayout {
 			model: 7
 			Day {
 				label: labels[index]
-				fg: {
-					if (index === weekdayOffset) return todayFg
-					return (index < weekdayOffset) ? pastDayFg : futureDayFg
-				}
-				bg: {
-					if (index === weekdayOffset) return todayBg
-					return (index < weekdayOffset) ? pastDayBg : futureDayBg
-				}
-				bold: {
-					if (index === weekdayOffset) return todayBold
-					return (index < weekdayOffset) ? pastDayBold : futureDayBold
-				}
-				italic: {
-					if (index === weekdayOffset) return todayItalic
-					return (index < weekdayOffset) ? pastDayItalic : futureDayItalic
-				}
+				fg: attrsFg[index]
+				bg: attrsBg[index]
+				bold: attrsBold[index]
+//				italic: attrsItalic[index]
 			}
 		}
 	} // weekGrid